# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payments/flouci/initiate": {
      "post": {
        "description": "Génère un lien de paiement Flouci pour une commande",
        "operationId": "FlouciController_initiatePayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paiement initié avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentInitiationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          },
          "404": {
            "description": "Commande non trouvée"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Initier un paiement",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/webhook": {
      "post": {
        "description": "Endpoint pour recevoir les notifications de paiement",
        "operationId": "FlouciController_handleWebhook",
        "parameters": [
          {
            "name": "x-flouci-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook traité",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDto"
                }
              }
            }
          }
        },
        "summary": "Webhook Flouci",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/verify": {
      "post": {
        "operationId": "FlouciController_verifyPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statut du paiement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentVerificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Vérifier le statut d'un paiement",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/status/{orderId}": {
      "get": {
        "operationId": "FlouciController_getPaymentStatus",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "ID de la commande",
            "schema": {
              "example": "order-uuid-123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statut du paiement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir le statut d'un paiement",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/cancel": {
      "post": {
        "description": "Annule un paiement en attente",
        "operationId": "FlouciController_cancelPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trackingId": {
                    "type": "string",
                    "description": "ID de tracking du paiement",
                    "example": "order-uuid-timestamp-random"
                  }
                },
                "required": [
                  "trackingId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paiement annulé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Annuler un paiement",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/test/simulate": {
      "post": {
        "description": "Utiliser 111111 pour succès, 000000 pour échec",
        "operationId": "FlouciController_simulateTestPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestWalletPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paiement simulé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentInitiationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Simuler un paiement de test (sandbox uniquement)",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/payments/flouci/admin/stats": {
      "get": {
        "operationId": "FlouciController_getPaymentStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistiques des paiements"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Statistiques de paiement (Admin)",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/admin/payments": {
      "get": {
        "operationId": "FlouciController_getPayments",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des paiements"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[ADMIN] Liste des paiements",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/admin/payments/{id}": {
      "get": {
        "operationId": "FlouciController_getPaymentDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détail du paiement"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[ADMIN] Détail dun paiement",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/admin/payments/{id}/refund": {
      "post": {
        "operationId": "FlouciController_refundPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remboursement effectué"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[ADMIN] Rembourser un paiement",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/admin/config": {
      "get": {
        "operationId": "FlouciController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configuration actuelle"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[ADMIN] Configuration Flouci",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/flouci/admin/trends": {
      "get": {
        "operationId": "FlouciController_getPaymentTrends",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tendance"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[ADMIN] Tendance des paiements",
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "API documentation for the application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "JWT-refresh": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Refresh-JWT",
        "description": "Enter refresh JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "InitiatePaymentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "ID de la commande",
            "example": "order-uuid-123"
          },
          "amount": {
            "type": "number",
            "description": "Montant en dinars tunisiens",
            "example": 99.99
          },
          "paymentReason": {
            "type": "string",
            "description": "Raison du paiement",
            "example": "Paiement commande #12345"
          },
          "customerEmail": {
            "type": "string",
            "description": "Email du client",
            "example": "client@example.com"
          },
          "customerPhone": {
            "type": "string",
            "description": "Numéro de téléphone du client",
            "example": "+216 98 765 432"
          },
          "returnUrl": {
            "type": "string",
            "description": "URL de retour personnalisée",
            "example": "http://localhost:3000/payment/return"
          },
          "metadata": {
            "type": "object",
            "description": "Métadonnées additionnelles",
            "example": {
              "cartId": "cart-uuid",
              "userId": "user-uuid"
            }
          }
        },
        "required": [
          "orderId",
          "amount"
        ]
      },
      "PaymentInitiationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Succès de l'opération",
            "example": true
          },
          "paymentId": {
            "type": "string",
            "description": "ID du paiement Flouci",
            "example": "flouci-payment-id"
          },
          "paymentLink": {
            "type": "string",
            "description": "Lien de redirection vers Flouci",
            "example": "https://sandbox.flouci.com/payment/xxx"
          },
          "trackingId": {
            "type": "string",
            "description": "ID de tracking",
            "example": "order-uuid-timestamp"
          },
          "amount": {
            "type": "number",
            "description": "Montant en dinars",
            "example": 99.99
          },
          "currency": {
            "type": "string",
            "description": "Devise",
            "example": "TND"
          },
          "expiresAt": {
            "type": "string",
            "description": "Date d'expiration",
            "example": "2024-01-01T13:00:00Z"
          },
          "message": {
            "type": "string",
            "description": "Message",
            "example": "Paiement initié avec succès"
          }
        },
        "required": [
          "success",
          "paymentId",
          "paymentLink",
          "trackingId",
          "amount",
          "currency",
          "expiresAt"
        ]
      },
      "WebhookResponseDto": {
        "type": "object",
        "properties": {
          "received": {
            "type": "boolean",
            "description": "Webhook reçu avec succès",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Message de confirmation",
            "example": "Webhook traité avec succès"
          },
          "orderId": {
            "type": "string",
            "description": "ID de la commande mise à jour",
            "example": "order-uuid-123"
          },
          "orderStatus": {
            "type": "string",
            "description": "Nouveau statut de la commande",
            "example": "PAID"
          }
        },
        "required": [
          "received",
          "message"
        ]
      },
      "VerifyPaymentDto": {
        "type": "object",
        "properties": {
          "developerTrackingId": {
            "type": "string",
            "description": "ID de tracking développeur",
            "example": "order-uuid-timestamp"
          },
          "paymentId": {
            "type": "string",
            "description": "ID du paiement Flouci",
            "example": "flouci-payment-id"
          }
        },
        "required": [
          "developerTrackingId"
        ]
      },
      "PaymentVerificationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Succès",
            "example": true
          },
          "paymentId": {
            "type": "string",
            "description": "ID du paiement",
            "example": "flouci-payment-id"
          },
          "trackingId": {
            "type": "string",
            "description": "ID de tracking",
            "example": "order-uuid-timestamp"
          },
          "status": {
            "type": "string",
            "description": "Statut du paiement",
            "enum": [
              "PENDING",
              "SUCCESS",
              "FAILED",
              "EXPIRED"
            ],
            "example": "SUCCESS"
          },
          "amount": {
            "type": "number",
            "description": "Montant en dinars",
            "example": 99.99
          },
          "currency": {
            "type": "string",
            "description": "Devise",
            "example": "TND"
          },
          "paidAt": {
            "type": "string",
            "description": "Date du paiement",
            "example": "2024-01-01T12:05:00Z"
          },
          "message": {
            "type": "string",
            "description": "Message d'erreur",
            "example": "Paiement effectué avec succès"
          }
        },
        "required": [
          "success",
          "paymentId",
          "trackingId",
          "status",
          "amount",
          "currency",
          "paidAt"
        ]
      },
      "PaymentStatusResponseDto": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "description": "ID du paiement",
            "example": "flouci-payment-id"
          },
          "trackingId": {
            "type": "string",
            "description": "ID de tracking",
            "example": "order-uuid-timestamp"
          },
          "orderId": {
            "type": "string",
            "description": "ID de la commande associée",
            "example": "order-uuid-123"
          },
          "amount": {
            "type": "number",
            "description": "Montant",
            "example": 99.99
          },
          "currency": {
            "type": "string",
            "description": "Devise",
            "example": "TND"
          },
          "status": {
            "type": "string",
            "description": "Statut",
            "enum": [
              "PENDING",
              "SUCCESS",
              "FAILED",
              "EXPIRED",
              "CANCELLED"
            ],
            "example": "SUCCESS"
          },
          "paymentReason": {
            "type": "string",
            "description": "Raison du paiement",
            "example": "Paiement commande #12345"
          },
          "createdAt": {
            "type": "string",
            "description": "Date de création",
            "example": "2024-01-01T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Date de mise à jour",
            "example": "2024-01-01T12:05:00Z"
          },
          "paidAt": {
            "type": "string",
            "description": "Date de paiement",
            "example": "2024-01-01T12:05:00Z"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Mode de paiement",
            "example": "CARD"
          },
          "metadata": {
            "type": "object",
            "description": "Métadonnées",
            "example": {
              "userId": "user-uuid"
            }
          }
        },
        "required": [
          "paymentId",
          "trackingId",
          "orderId",
          "amount",
          "currency",
          "status",
          "paymentReason",
          "createdAt",
          "updatedAt"
        ]
      },
      "TestWalletPaymentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "ID de la commande",
            "example": "order-uuid-123"
          },
          "amount": {
            "type": "number",
            "description": "Montant en dinars",
            "example": 10
          },
          "testWalletNumber": {
            "type": "string",
            "description": "Numéro de wallet de test (111111 = succès, 000000 = échec)",
            "example": "111111"
          },
          "paymentReason": {
            "type": "string",
            "description": "Raison du paiement",
            "example": "Test paiement"
          }
        },
        "required": [
          "orderId",
          "amount",
          "testWalletNumber"
        ]
      },
...
    }
  }
}